pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-timeline-extractor"
version = "0.1.0"
description = "YouTube or local video → structured event timeline (transcript + LLM)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "faster-whisper>=1.0.0",
    "yt-dlp>=2024.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
]
# Speaker diarization (heavy: PyTorch + model downloads; requires HF_TOKEN and model license acceptance).
diarize = [
    "pyannote.audio>=3.1.0",
]

[project.scripts]
ete = "event_timeline_extractor.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
event_timeline_extractor = ["web/static/*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"src/event_timeline_extractor/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: downloads models or heavy ASR",
    "network: requires outbound network",
]
addopts = "-q --import-mode=importlib"
